# docs/swagger.yaml
basePath: /api
definitions:
...
  handlers.BulkCreateResult:
    properties:
      email:
        type: string
      error:
        type: string
      index:
        type: integer
      success:
        type: boolean
      user_id:
        type: string
    type: object
  handlers.BulkCreateUsersRequest:
    properties:
      users:
        items:
          $ref: '#/definitions/models.UserCreateRequest'
        type: array
    required:
    - users
    type: object
  handlers.BulkCreateUsersResponse:
    properties:
      error_count:
        type: integer
      results:
        items:
          $ref: '#/definitions/handlers.BulkCreateResult'
        type: array
      success_count:
        type: integer
      total_processed:
        type: integer
    type: object
...
  models.ErrorResponse:
    properties:
      code:
        example: 401
        type: integer
      details: {}
      error:
        example: Invalid credentials
        type: string
      message:
        example: The provided email or password is incorrect
        type: string
    type: object
  models.LogData:
    properties:
      action:
        type: string
      details:
        additionalProperties: true
        type: object
      duration:
        description: Request duration in milliseconds
        type: integer
      error:
        description: For error events
        type: string
      new_values:
        additionalProperties: true
        description: For update operations
        type: object
      old_values:
        additionalProperties: true
        description: For update operations
        type: object
      status_code:
        description: HTTP status code
        type: integer
    type: object
  models.LogEventType:
    enum:
    - USER_CREATED
    - USER_UPDATED
    - USER_DELETED
    - USER_LOGIN
    - ADMIN_LOGIN
    - ADMIN_LOGOUT
    - LOGIN_SUCCESS
    - LOGIN_FAILED
    - TOKEN_REFRESH
    - SYSTEM_ERROR
    - VALIDATION_ERROR
    type: string
    x-enum-varnames:
    - UserCreated
    - UserUpdated
    - UserDeleted
    - UserLogin
    - AdminLogin
    - AdminLogout
    - LoginSuccess
    - LoginFailed
    - TokenRefresh
    - SystemError
    - ValidationLogError
...
  models.SuccessResponse:
    properties:
      data: {}
      message:
        example: Operation completed successfully
        type: string
      success:
        example: true
        type: boolean
    type: object
  models.UserCreateRequest:
    properties:
      email:
        example: john.doe@example.com
        type: string
      name:
        example: John Doe
        type: string
      password:
        example: password123
        minLength: 6
        type: string
    required:
    - email
    - name
    - password
    type: object
  models.UserLogResponse:
    properties:
      data:
        $ref: '#/definitions/models.LogData'
      event:
        $ref: '#/definitions/models.LogEventType'
      id:
        type: string
      ip_address:
        type: string
      timestamp:
        type: string
      user_agent:
        type: string
      user_id:
        type: string
    type: object
  models.UserLogsListResponse:
    properties:
      logs:
        items:
          $ref: '#/definitions/models.UserLogResponse'
        type: array
      page:
        type: integer
      page_size:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  models.UserResponse:
    properties:
      created_at:
        example: "2023-01-01T00:00:00Z"
        type: string
      email:
        example: john.doe@example.com
        type: string
      id:
        example: 550e8400-e29b-41d4-a716-446655440000
        type: string
      name:
        example: John Doe
        type: string
      updated_at:
        example: "2023-01-01T00:00:00Z"
        type: string
    type: object
...
  models.UsersListResponse:
    properties:
      page:
        example: 1
        type: integer
      page_size:
        example: 10
        type: integer
      total:
        example: 100
        type: integer
      total_pages:
        example: 10
        type: integer
      users:
        items:
          $ref: '#/definitions/models.UserResponse'
        type: array
    type: object
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: A comprehensive user management system with JWT authentication, role-based
    access control, and activity logging.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: User Management System API
  version: "1.0"
paths:
  /api/admin/logs:
    get:
      consumes:
      - application/json
      description: Get paginated user activity logs with filtering options
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      - description: Filter by user ID
        in: query
        name: user_id
        type: string
      - description: Filter by event type
        in: query
        name: event
        type: string
      - description: Start date (RFC3339)
        in: query
        name: start_date
        type: string
      - description: End date (RFC3339)
        in: query
        name: end_date
        type: string
      - description: Filter by IP address
        in: query
        name: ip_address
        type: string
      - description: Filter by action
        in: query
        name: action
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UserLogsListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user activity logs
      tags:
      - admin
  /api/admin/maintenance:
    post:
      description: Run system maintenance tasks (log cleanup, etc.)
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Run system maintenance
      tags:
      - admin
  /api/admin/stats:
    get:
      description: Get comprehensive system statistics and metrics
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get system statistics
      tags:
      - admin
  /api/admin/users/{id}/permanent-delete:
    delete:
      description: Permanently delete a user from the database (irreversible)
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Permanently delete user
      tags:
      - admin
  /api/admin/users/{id}/restore:
    post:
      description: Restore a soft-deleted user account
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restore deleted user
      tags:
      - admin
  /api/admin/users/bulk-create:
    post:
      consumes:
      - application/json
      description: Create multiple users in a single operation
      parameters:
      - description: Bulk user creation data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkCreateUsersRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/handlers.BulkCreateUsersResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Bulk create users
      tags:
      - admin
  /api/admin/users/deleted:
    get:
      consumes:
      - application/json
      description: Get list of soft-deleted users for potential restoration
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Page size
        in: query
        name: page_size
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UsersListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get deleted users
      tags:
      - admin
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...